(fixed) split test_datasaver_filetype to torch and tensorflow version
